requirements.txt
requests
redis
spacy
python-dotenv
geopy
